[engine][server] Refactoring. Moved RoomService to engine.
