+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/ScoresCleanerTest.java
@@ -26,8 +26,8 @@ import com.codenjoy.dojo.config.Constants;
 import com.codenjoy.dojo.services.dao.Registration;
 import com.codenjoy.dojo.services.helper.Helpers;
 import com.codenjoy.dojo.services.info.Information;
-import com.codenjoy.dojo.services.mocks.FakePlayerScores;
-import com.codenjoy.dojo.services.mocks.FirstGameType;
+import com.codenjoy.dojo.utils.test.mocks.FakePlayerScores;
+import com.codenjoy.dojo.utils.test.mocks.FirstGameType;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
