+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/mocks/FirstSemifinalGameType.java
@@ -24,6 +24,7 @@ package com.codenjoy.dojo.services.mocks;
 
 
 import com.codenjoy.dojo.services.settings.SettingsImpl;
+import com.codenjoy.dojo.utils.test.mocks.FirstGameType;
 
 public class FirstSemifinalGameType extends FirstGameType {
