+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/mocks/SecondSemifinalGameSettings.java
@@ -23,6 +23,7 @@ package com.codenjoy.dojo.services.mocks;
  */
 
 import com.codenjoy.dojo.services.semifinal.SemifinalSettings;
+import com.codenjoy.dojo.utils.test.mocks.SecondGameSettings;
 
 public class SecondSemifinalGameSettings extends SecondGameSettings
         implements SemifinalSettings<SecondGameSettings> {
