+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/mocks/SecondSemifinalGameType.java
@@ -24,6 +24,7 @@ package com.codenjoy.dojo.services.mocks;
 
 
 import com.codenjoy.dojo.services.settings.SettingsImpl;
+import com.codenjoy.dojo.utils.test.mocks.SecondGameType;
 
 public class SecondSemifinalGameType extends SecondGameType {
